api/requirements.txt
Flask-SQLAlchemy
psycopg2-binary
Flask-Migrate
python-dotenv
Flask-JWT-Extended
Marshmallow
Marshmallow-SQLAlchemy
Werkzeug
Flask-Cors